unit test unstructured_helper.go 3
