+++ pkg/kube/unstructured/test-fixtures/files/analysis-template.yaml
