+++ pkg/kube/unstructured/test-fixtures/files/instance-group-not-ready.yaml
