+++ pkg/kube/unstructured/test-fixtures/files/instance-group.yaml
