+++ pkg/kube/unstructured/test-fixtures/files/multi-resource-no-ns.yaml
