+++ pkg/kube/unstructured/test-fixtures/files/multi-resource.yaml
