+++ pkg/kube/unstructured/test-fixtures/files/resource-no-ns.yaml
