+++ pkg/kube/unstructured/test-fixtures/files/resource.yaml
