+++ pkg/kube/unstructured/test-fixtures/templates/templated.yaml
